+++ src/lumora.h
@@ -11,6 +11,7 @@ std::string readFile(const char* path);
 void pushArgs(lua_State* L, int argc, char** argv, int scriptIndex);
 void applySandbox(lua_State* L);
 void registerCLIGlobals(lua_State* L);
+void registerRequire(lua_State* L);
 void freezeLibraries(lua_State* L);
 int runScript(const char* path, int argc, char** argv, bool roblox, bool sandbox, double timeout);
 int runVisual(const char* path, int argc, char** argv, bool sandbox);